runtime/consensus/roothash: Reorganize the roothash module
The roothash module was split into multiple files because having
all the code in one file became overwhelming.
