+++ runtime/src/consensus/roothash/commitment/mod.rs
@@ -1,0 +1,4 @@
+mod executor;
+
+// Re-exports.
+pub use executor::*;
